requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.2
django-auto-prefetching==0.2.12
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
